Adds the possibility to provide a specific schema for the flyway meta data table.
Flyway will automatically add the flyway meta data table (flyway_schema_history) to the first schema provided. Therefore jOOQ will also generate a class for it, which might (or should) not be needed in the production code. This change adds the possibility to provide a specific schema for the meta data table which will not be fed into the jOOQ class generation.