Removes tty and stdin configuration from await container start command.
On one hand, the `withTty(true)` configuration is removed since it is probably just not needed (anymore), on the other hand, `withStdIn...` leads to a problem. We experience the following issue (with `docker-java`) described [here](https://github.com/docker-java/docker-java/issues/1350) (or related) on MacOS systems. Every container run leaves a Java process behind that uses 100% of a CPU core and never terminates. This is pretty annoying since these processes need to be killed manually. On Linux boxes we could not observe this problem. Since the issue has not been fixed in `docker-java` and the attached std in is not vital for the the container start wait command to function properly, it is removed.